Update importer to work with newer blender SCM
So the blender exporter can do dirty tricks the Max importer cant deal
with, and so the import script is now more robust to make that work. Old
blender-mad SCM still dont work since the new blender tool got a
compatibility update that changes the padding character.

someday, someone will need to update the import plugin so that it doesnt
look for a hardcoded padding character, but that wont be me, probably.